requirements.txt
allure-pytest==2.8.13
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
openpyxl==3.0.3
pluggy==0.13.1
pyparsing==2.4.7
pytest==5.4.1
pytest-assume==2.2.1
pytest-timeout==1.3.4
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
vertica-python==0.11.0
wcwidth==0.1.9
zipp==3.1.0
